medical-app-backed-stage02: Refactored controllers implementation.

+++ medical-app-backend/src/main/java/com/fral/medapp/controller/MedicalExamRestController.java
@@ -1,7 +1,9 @@
 package com.fral.medapp.controller;
 
-import java.util.ArrayList;
+import java.net.URI;
 import java.util.List;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,7 +17,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.fral.medapp.exception.EntityNotFoundException;
 import com.fral.medapp.model.MedicalExam;
 import com.fral.medapp.service.MedicalExamService;
 
@@ -28,54 +32,45 @@ public class MedicalExamRestController {
 
 	@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<List<MedicalExam>> listar() {
-		List<MedicalExam> examenes = new ArrayList<>();
-		examenes = service.getAll();
+		List<MedicalExam> examenes = service.getAll();
+		
 		return new ResponseEntity<List<MedicalExam>>(examenes, HttpStatus.OK);
 	}
 	
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<MedicalExam> listarId(@PathVariable("id") Integer id) {
-		MedicalExam examen = new MedicalExam();
-		examen = service.findById(id);
+		MedicalExam examen = service.findById(id);
+		
+		if (examen == null) {
+			throw new EntityNotFoundException("Id: " + id);
+		}
 		return new ResponseEntity<MedicalExam>(examen, HttpStatus.OK);
 	}
 
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> registrar(@RequestBody MedicalExam examen) {
-		int resultado = 0;
-		try {
-			service.create(examen);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
-		}
+	public ResponseEntity<Integer> registrar(@Valid @RequestBody MedicalExam examen) {
+		service.create(examen);
+		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
+					   .buildAndExpand(examen.getId()).toUri();
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+		return ResponseEntity.created(location).build();
 	}
 
 	@PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> actualizar(@RequestBody MedicalExam examen) {
-		int resultado = 0;
-		try {
-			service.update(examen);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
-		}
+	public ResponseEntity<Object> actualizar(@Valid @RequestBody MedicalExam examen) {
+		service.update(examen);
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 	@DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> eliminar(@PathVariable Integer id) {
-		int resultado = 0;
-		try {
-			service.delete(id);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
+	public void eliminar(@PathVariable Integer id) {
+		MedicalExam medicalExam = service.findById(id);
+		
+		if (medicalExam == null) {
+			throw new EntityNotFoundException("Id: " + id);
 		}
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+		service.delete(id);
 	}
 }
