+++ medical-app-backend/src/main/java/com/fral/medapp/controller/SpecialtyRestController.java
@@ -1,7 +1,10 @@
 package com.fral.medapp.controller;
 
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,7 +18,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.fral.medapp.exception.EntityNotFoundException;
 import com.fral.medapp.model.Specialty;
 import com.fral.medapp.service.SpecialtyService;
 
@@ -35,48 +40,42 @@ public class SpecialtyRestController {
 	
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Specialty> listarId(@PathVariable("id") Integer id) {
-		Specialty especialidad = new Specialty();
-		especialidad = service.findById(id);
+		Specialty especialidad = service.findById(id);
+
+		if (especialidad == null) {
+			throw new EntityNotFoundException("Id: " + id);
+		}
 		return new ResponseEntity<Specialty>(especialidad, HttpStatus.OK);
 	}
 
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> registrar(@RequestBody Specialty especialidad) {
-		int resultado = 0;
-		try {
-			service.create(especialidad);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
-		}
+	public ResponseEntity<Object> registrar(@Valid @RequestBody Specialty especialidad) {
+		
+		service.create(especialidad);
+		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
+				       .buildAndExpand(especialidad.getId()).toUri();
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+		return ResponseEntity.created(location).build();
 	}
 
 	@PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> actualizar(@RequestBody Specialty especialidad) {
-		int resultado = 0;
-		try {
-			service.update(especialidad);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
-		}
+	public ResponseEntity<Object> actualizar(@Valid @RequestBody Specialty especialidad) {
+		
+		service.update(especialidad);
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 	@DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> eliminar(@PathVariable Integer id) {
-		int resultado = 0;
-		try {
-			service.delete(id);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
-		}
+	public void eliminar(@PathVariable Integer id) {
+		
+		Specialty specialty = service.findById(id);
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+		if (specialty == null) {
+			throw new EntityNotFoundException("Id: " + id);
+		}
+		
+		service.delete(id);
 	}
 
 }
